+++ lib/src/main/scala/com/github/gchudnov/kprojekt/files/FileOps.scala
@@ -1,4 +1,4 @@
-package com.github.gchudnov.files
+package com.github.gchudnov.kprojekt.files
 
 import java.io.BufferedWriter
 import java.io.File
