+++ lib/src/main/scala/com/github/gchudnov/kprojekt/parser/ParseException.scala
@@ -1,3 +1,3 @@
-package com.github.gchudnov.parser
+package com.github.gchudnov.kprojekt.parser
 
 final class ParseException(message: String) extends RuntimeException(message)
