+++ lib/src/main/scala/com/github/gchudnov/kprojekt/parser/blocks/ProcessorBlock.scala
@@ -1,4 +1,4 @@
-package com.github.gchudnov.parser.blocks
+package com.github.gchudnov.kprojekt.parser.blocks
 
 import java.util.{Set => JSet}
 import org.apache.kafka.streams.TopologyDescription.Processor
