+++ lib/src/main/scala/com/github/gchudnov/kprojekt/parser/blocks/TopologyDescriptionBlock.scala
@@ -1,4 +1,4 @@
-package com.github.gchudnov.parser.blocks
+package com.github.gchudnov.kprojekt.parser.blocks
 
 import java.util.{Set => JSet, HashSet => JHashSet}
 import org.apache.kafka.streams.TopologyDescription
