+++ lib/src/main/scala/com/github/gchudnov/kprojekt/presenter/render/GraphEasyRender.scala
@@ -1,0 +1,1 @@
+package com.github.gchudnov.kprojekt.presenter.render
