+++ lib/src/main/scala/com/github/gchudnov/kprojekt/presenter/render/NodeName.scala
@@ -1,4 +1,4 @@
-package com.github.gchudnov.presenter.name
+package com.github.gchudnov.kprojekt.presenter.render
 
 import scala.util.matching.Regex
