--- lib/src/main/scala/com/github/gchudnov/presenter/render/GraphEasyRender.scala
@@ -1,1 +1,0 @@
-package com.github.gchudnov.presenter.render
