--- lib/src/test/scala/com/github/gchudnov/files/FileOps.scala
@@ -1,16 +1,0 @@
-package com.github.gchudnov.files
-
-import java.io.BufferedWriter
-import java.io.File
-import java.io.FileWriter
-import scala.util.Using
-
-object FileOps {
-
-  def save(file: File)(data: String) = {
-    Using(new BufferedWriter(new FileWriter(file))) { writer =>
-      writer.write(data)
-    }
-  }
-
-}
