+++ lib/src/test/scala/com/github/gchudnov/kprojekt/presenter/render/DotRenderSpec.scala
@@ -1,4 +1,4 @@
-package com.github.gchudnov.presenter.render
+package com.github.gchudnov.kprojekt.presenter.render
 
 import org.scalatest.{WordSpec, Matchers}
