+++ project/plugins.sbt
@@ -1,5 +1,6 @@
 addSbtPlugin("com.timushev.sbt" % "sbt-updates" % "0.4.2")
 addSbtPlugin("com.eed3si9n" % "sbt-assembly" % "0.14.10")
+addSbtPlugin("com.eed3si9n" % "sbt-buildinfo" % "0.9.0")
 addSbtPlugin("com.typesafe.sbt" % "sbt-native-packager" % "1.4.1")
 
 //addSbtPlugin("com.github.gseitz" % "sbt-release" % "1.0.8")
